requirements.txt
Django~=4.2
psycopg2-binary
google-generativeai
python-dotenv
dj-database-url